@@ -14,50 +14,49 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/cdk": "^15.2.1",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/material": "^15.2.9",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/youtube-player": "^14.2.0",
+    "@angular/animations": "^18.2.5",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.2.5",
+    "@angular/compiler": "^18.2.5",
+    "@angular/core": "^18.2.5",
+    "@angular/forms": "^18.2.5",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.2.5",
+    "@angular/platform-browser-dynamic": "^18.2.5",
+    "@angular/router": "^18.2.5",
+    "@angular/youtube-player": "^18.2.4",
     "@kolkov/angular-editor": "^3.0.0-beta.0",
-    "@ng-idle/core": "^15.0.0",
-    "@ng-idle/keepalive": "^15.0.0",
-    "angular2-moment": "^1.9.0",
+    "@ng-idle/keepalive": "^16.0.0",
     "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.11.1",
     "jquery": "^3.6.1",
-    "ngx-bootstrap": "^10.0.0",
+    "ngx-bootstrap": "^18.0.2",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "typescript": "^5.1.6",
     "zone.js": "~0.14.10"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^18.2.5",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "^17.3.9",
-    "@angular/compiler-cli": "^17.3.12",
+    "@angular-eslint/builder": "^18.0.0",
+    "@angular-eslint/eslint-plugin": "^18.0.0",
+    "@angular-eslint/eslint-plugin-template": "^18.0.0",
+    "@angular-eslint/schematics": "^18.0.0",
+    "@angular-eslint/template-parser": "^18.0.0",
+    "@angular/cli": "^18.2.5",
+    "@angular/compiler-cli": "^18.2.5",
     "@cypress/schematic": "^2.5.1",
+    "@ng-idle/core": "^16.0.0",
     "@types/jasmine": "~4.3.0",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
+    "@typescript-eslint/eslint-plugin": "7.11.0",
+    "@typescript-eslint/parser": "7.11.0",
     "cypress": "latest",
-    "eslint": "^8.33.0",
+    "eslint": "8.57.0",
     "jasmine-core": "~4.5.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0"
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "typescript": "^5.5.4"
   }
 }
