@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { EmployeeService } from './httpServ/employee.service';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { DynamicInterceptorInterceptor } from './dynamic-interceptor.interceptor';
 import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 //import { TimeoutModalComponent } from './timeout-modal/timeout-modal.component';
@@ -11,24 +11,15 @@ import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 //import { ModalModule } from 'ngx-bootstrap/modal';
 //import { HashLocationStrategy, LocationStrategy } from '@angular/common';
 
-@NgModule({
-  declarations: [
+@NgModule({ declarations: [
     //TimeoutModalComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    //AppRoutingModule
-    //NgIdleKeepaliveModule.forRoot(),
-    //ModalModule.forRoot(),
-  ],
-  providers: [
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: DynamicInterceptorInterceptor,
-      multi: true,
-    },
-  ],
-  bootstrap: [],
-})
+    ],
+    bootstrap: [], imports: [BrowserModule], providers: [
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: DynamicInterceptorInterceptor,
+            multi: true,
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
